+++ tasks/clustering.py
@@ -48,7 +48,8 @@ logger = logging.getLogger(__name__)
 
 def batch_task_failure_handler(job, connection, type, value, tb):
     """A failure handler for the clustering batch sub-task, executed by the worker."""
-    from app import app, save_task_status, TASK_STATUS_FAILURE
+    from app import app
+    from app_helper import save_task_status, TASK_STATUS_FAILURE
     with app.app_context():
         task_id = job.get_id()
         parent_id = job.kwargs.get('parent_task_id')
@@ -126,7 +127,8 @@ def run_clustering_batch_task(
     Executes a batch of clustering iterations. This task is enqueued by the main clustering task.
     """
     # --- Local imports to prevent circular dependency ---
-    from app import (app, redis_conn, save_task_status, get_task_info_from_db,
+    from app import app
+    from app_helper import (redis_conn, save_task_status, get_task_info_from_db,
                      TASK_STATUS_PROGRESS, TASK_STATUS_REVOKED, TASK_STATUS_FAILURE,
                      TASK_STATUS_SUCCESS)
 
@@ -265,11 +267,11 @@ def run_clustering_task(
     Orchestrates data preparation, batch job creation, result aggregation, and playlist creation.
     """
     # --- Local imports to prevent circular dependency ---
-    from app import (app, redis_conn, get_db, save_task_status, get_task_info_from_db,
-                     update_playlist_table,
+    from app import app
+    from app_helper import (redis_conn, get_db, save_task_status, get_task_info_from_db,
+                     update_playlist_table, get_child_tasks_from_db,
                      TASK_STATUS_PENDING, TASK_STATUS_STARTED, TASK_STATUS_PROGRESS,
-                     TASK_STATUS_SUCCESS, TASK_STATUS_FAILURE, TASK_STATUS_REVOKED,
-                     get_child_tasks_from_db)
+                     TASK_STATUS_SUCCESS, TASK_STATUS_FAILURE, TASK_STATUS_REVOKED)
 
     current_job = get_current_job(redis_conn)
     current_task_id = current_job.id if current_job else str(uuid.uuid4())
@@ -536,9 +538,9 @@ def _monitor_and_process_batches(state_dict, parent_task_id, initial_check=False
     database but is missed by the parent due to a race condition with RQ job cleanup.
     """
     # --- Local import to prevent circular dependency ---
-    from app import get_child_tasks_from_db, get_task_info_from_db, redis_conn, \
+    from app_helper import (redis_conn, get_child_tasks_from_db, get_task_info_from_db,
                     TASK_STATUS_SUCCESS, TASK_STATUS_FAILURE, TASK_STATUS_REVOKED, \
-                    TASK_STATUS_PENDING, TASK_STATUS_STARTED, TASK_STATUS_PROGRESS
+                    TASK_STATUS_PENDING, TASK_STATUS_STARTED, TASK_STATUS_PROGRESS)
     
     # 1. Get all child tasks from the database.
     all_child_tasks = get_child_tasks_from_db(parent_task_id)
@@ -647,7 +649,7 @@ def _monitor_and_process_batches(state_dict, parent_task_id, initial_check=False
 
 def _launch_batch_job(state_dict, parent_task_id, batch_idx, total_runs, genre_map, target_per_genre, *args):
     """Constructs and enqueues a single batch job."""
-    from app import rq_queue_default # Local import to avoid circular dependency issues at top-level
+    from app_helper import rq_queue_default # Local import to avoid circular dependency issues at top-level
 
     # Unpack all the parameters passed via *args
     (
